Add reply method and improved tests for RedditService.
This method gets a comment object (with PRAW modelling) and a list of medal objects
and replies to it with the correct response text. It can also handle the cases where
there arent any medals matching the requirements.

We've also added its tests. They are still empty, but the fact that they are
already defined will give us a backbone to extend them later