Modifying BaseDBTestCase so it can be used in our entire test suite
We've extracted it to an independent file, test/helpers.py, to make it clearer that it is intended to be used in our entire test base and not only on the models tests.

We've also modified its setUp method to stop binding the models there. Instead we do it now on the run method, by using the bind_ctx context manager.